fix bug

+++ paddleslim/quant/observers/abs_max_weight.py
@@ -65,8 +65,7 @@ class AbsMaxChannelWiseWeightObserverLayer(ChannelWiseObserver):
     def _cal_abs_max(self, inputs):
         reduce_axis = tuple(
             [i for i in range(len(inputs.shape)) if i != self.quant_axis()])
-        abs_max_values = paddle.max(
-            paddle.abs(inputs), axis=reduce_axis).cast("float32")
+        abs_max_values = paddle.max(paddle.abs(inputs), axis=reduce_axis)
         abs_max_values = paddle.where(
             abs_max_values == paddle.to_tensor(0, dtype=inputs.dtype),
             paddle.to_tensor(1e-8, dtype=inputs.dtype), abs_max_values)
